Remove the tonatona-db module and add Config and HasConfig to tonatona-db-* child modules.
