+++ tonatona-db-postgresql/tonatona-db-postgresql.cabal
@@ -26,7 +26,6 @@ library
                      , persistent-postgresql
                      , resource-pool
                      , tonatona
-                     , tonatona-db
                      , tonatona-environment
                      , unliftio
   default-language:    Haskell2010
