+++ tonatona-db-sqlite/tonatona-db-sqlite.cabal
@@ -27,7 +27,6 @@ library
                      , resource-pool
                      , text
                      , tonatona
-                     , tonatona-db
                      , unliftio
   default-language:    Haskell2010
   ghc-options:         -Wall -fwarn-incomplete-uni-patterns -fwarn-incomplete-record-updates -fwarn-monomorphism-restriction
