--- tonatona-db/README.md
@@ -1,3 +1,0 @@
-# tonatona-db
-
-A plugin for `tonatona` that adds features to access db.
