--- tonatona-db/test/Spec.hs
@@ -1,6 +1,0 @@
-{-# LANGUAGE QuasiQuotes #-}
-
-module Main where
-
-main :: IO ()
-main = pure ()
